leerHugging/requirements.txt
datasets
pandas
mysql-connector-python
